requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
reportlab>=4.0.0
Pillow>=10.0.0
pandas>=2.1.0
